Refactor article commands and enhance document previews
- Changed data types for `DefaultQuantity` and `MinimumStock` in `AddArticleCommand` and `UpdateArticleCommand` from `uint` to `decimal` for better precision.
- Introduced `GetDocumentPreviewPdfQuery` and `GetDocumentPreviewPdfQueryHandler` for handling document preview requests with new properties.
- Updated parsing in `ArticleModal` and `UpdateArticleModal` to align with new data types.
- Enhanced `EntitySelectionModal` with event handlers for improved user interaction.
- Added validation checks in `CreateDocumentModal` before generating previews.
- Expanded `DocumentPdfUtils` to include methods for document previewing.
- Modified `DocumentsPage` and `ReceiptsPage` to handle selected entity confirmations for direct document viewing.

+++ LogicPOS.Api/Features/Articles/Articles/AddArticle/AddArticleCommand.cs
@@ -17,9 +17,9 @@ namespace LogicPOS.Api.Features.Articles.AddArticle
         public ArticlePrice Price5 { get; set; }
         public bool PriceWithVat { get; set; }
         public decimal Discount { get; set; }
-        public uint DefaultQuantity { get; set; }
+        public decimal DefaultQuantity { get; set; }
         public decimal TotalStock { get; set; }
-        public uint MinimumStock { get; set; }
+        public decimal MinimumStock { get; set; }
         public decimal Tare { get; set; }
         public float Weight { get; set; }
         public string Barcode { get; set; }
